pin-1775: Split agreement operations spec
